requirements.txt
selenium==4.16.0
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
lxml
